SCC: rename style.py file to caption_style.py
