+++ src/main/python/ttconv/scc/caption_style.py
